+++ plugins/input/docker/stdout/docker_stdout_processor_test.go
@@ -23,9 +23,11 @@ import (
 	"time"
 
 	"github.com/pingcap/check"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/alibaba/ilogtail/helper"
+	"github.com/alibaba/ilogtail/pkg/logger"
 	"github.com/alibaba/ilogtail/pkg/protocol"
 	"github.com/alibaba/ilogtail/pkg/util"
 )
@@ -33,20 +35,28 @@ import (
 // go test -check.f "inputProcessorTestSuite.*" -args -debug=true -console=true
 var _ = check.Suite(&inputProcessorTestSuite{})
 
-var log1 = `{"log":"1:M 09 Nov 13:27:36.276 # User requested shutdown...\n","stream":"stdout", "time":"2018-05-16T06:28:41.2195434Z"
-`
+func Test(t *testing.T) {
+	logger.InitTestLogger()
+	check.TestingT(t)
+}
 
 // error json
-var context1 = `{"log":"1:M 09 Nov 13:27:36.276 # User requested shutdown...\n","stream":"stdout", "time":"2018-05-16T06:28:41.2195434Z"`
+var context1 = `1:M 09 Nov 13:27:36.276 # User requested shutdown...`
+var log1 = `{"log":"1:M 09 Nov 13:27:36.276 # User requested shutdown...\n","stream":"stdout", "time":"2018-05-16T06:28:41.2195434Z"}
+`
+
+var context2 = `1:M 09 Nov 13:27:36.276 # User requested begin...`
 var log2 = `{"log":"1:M 09 Nov 13:27:36.276 # User requested begin...\n","stream":"stderr", "time":"2018-05-16T06:28:41.2195434Z"}
 `
-var context2 = `1:M 09 Nov 13:27:36.276 # User requested begin...`
+
+var context3 = `2017-09-12 22:32:21.212 [INFO][88] table.go 710: Invalidating dataplane cache`
 var log3 = `2017-09-12T22:32:21.212861448Z stdout 2017-09-12 22:32:21.212 [INFO][88] table.go 710: Invalidating dataplane cache
 `
-var context3 = `2017-09-12 22:32:21.212 [INFO][88] table.go 710: Invalidating dataplane cache`
+
+var context4 = `2017-09-12 22:32:21.212 [INFO][88] table.go 710:`
 var log4 = `2017-09-12T22:32:21.212861448Z stderr 2017-09-12 22:32:21.212 [INFO][88] table.go 710:
 `
-var context4 = `2017-09-12 22:32:21.212 [INFO][88] table.go 710:`
+
 var log5 = `2017-09-12T22:32:21.212861448Z stderr 2017-09-12 22:32:21.212 [INFO][88] exception 1: 
 `
 var log6 = `2017-09-12T22:32:21.212861448Z stderr    java.io.xxx.exception1....
@@ -67,6 +77,10 @@ var context6 = `2017-09-12 22:32:21.212 [INFO][88] exception 2:
    java.io.xxx.exception22....`
 
 var logErrorJSON = `{"log":"1:M 09 Nov 13:27:36.276 # User requested shutdown...\n","stream":"stdout", "time":"2018-05-16T06:28:41.2195434Z"`
+
+var logErrorContainerd1 = `2017-09-12T22:32:21.212861448Zstderr2017-09-1222:32:21.212[INFO][88]exception2:`
+var logErrorContainerd2 = `2017-09-12T22:32:21.212861448Z stderr2017-09-1222:32:21.212[INFO][88]exception2:`
+var logErrorContainerd3 = `2017-09-12T22:32:21.212861448Z stderr F2017-09-1222:32:21.212[INFO][88]exception2:`
 
 var splitedlog1 = `{"log":"1","stream":"stdout", "time":"2018-05-16T06:28:41.2195434Z"}
 `
@@ -149,8 +163,9 @@ func (s *inputProcessorTestSuite) TearDownTest(c *check.C) {
 
 func (s *inputProcessorTestSuite) TestNormal(c *check.C) {
 	processor := NewDockerStdoutProcessor(nil, time.Duration(0), 0, 512*1024, true, true, &s.context, &s.collector, s.tag)
-
-	n := processor.Process([]byte(s.allLog), time.Duration(0))
+	bytes := []byte(s.allLog)
+	str := helper.ZeroCopyString(bytes)
+	n := processor.Process(bytes, time.Duration(0))
 	c.Assert(n, check.Equals, len(s.allLog))
 	c.Assert(len(s.collector.Logs), check.Equals, len(s.allLogContent))
 	for index, content := range s.normalContent {
@@ -163,11 +178,13 @@ func (s *inputProcessorTestSuite) TestNormal(c *check.C) {
 		} else {
 			c.Assert(s.collector.Logs[index].Contents[2].GetValue(), check.Equals, "stderr")
 		}
+		value := s.collector.Logs[index].Contents[1].GetValue()
 		if index < 2 {
-			c.Assert(s.collector.Logs[index].Contents[1].GetValue(), check.Equals, "2018-05-16T06:28:41.2195434Z")
+			c.Assert(value, check.Equals, "2018-05-16T06:28:41.2195434Z")
 		} else {
-			c.Assert(s.collector.Logs[index].Contents[1].GetValue(), check.Equals, "2017-09-12T22:32:21.212861448Z")
+			c.Assert(value, check.Equals, "2017-09-12T22:32:21.212861448Z")
 		}
+		c.Assert(helper.IsSafeString(str, value), check.IsTrue)
 		switch s.collector.Logs[index].Contents[3].GetKey() {
 		case "container_id":
 			c.Assert(s.collector.Logs[index].Contents[3].GetKey(), check.Equals, "container_id")
@@ -185,94 +202,211 @@ func (s *inputProcessorTestSuite) TestNormal(c *check.C) {
 
 func (s *inputProcessorTestSuite) TestSplitedLine(c *check.C) {
 	processor := NewDockerStdoutProcessor(nil, time.Second, 0, 512*1024, true, true, &s.context, &s.collector, s.tag)
-	n1 := processor.Process([]byte(splitedlog1), time.Duration(0))
-	n2 := processor.Process([]byte(splitedlog2), time.Duration(0))
-	n3 := processor.Process([]byte(splitedlog3), time.Duration(0))
+	splitedlog1Bytes := []byte(splitedlog1)
+	str1 := helper.ZeroCopyString(splitedlog1Bytes)
+	splited2og1Bytes := []byte(splitedlog2)
+	str2 := helper.ZeroCopyString(splited2og1Bytes)
+	splited3og1Bytes := []byte(splitedlog3)
+	str3 := helper.ZeroCopyString(splited3og1Bytes)
+
+	n1 := processor.Process(splitedlog1Bytes, time.Duration(0))
+	c.Assert(helper.IsSafeString(str1, helper.ZeroCopyString(processor.lastLogs[0].Content)), check.IsTrue)
 	c.Assert(n1, check.Equals, len(splitedlog1))
+
+	n2 := processor.Process(splited2og1Bytes, time.Duration(0))
+	c.Assert(helper.IsSafeString(str2, helper.ZeroCopyString(processor.lastLogs[1].Content)), check.IsTrue)
 	c.Assert(n2, check.Equals, len(splitedlog2))
+
+	n3 := processor.Process(splited3og1Bytes, time.Duration(0))
 	c.Assert(n3, check.Equals, len(splitedlog3))
+
 	c.Assert(len(s.collector.Logs), check.Equals, 1)
-	fmt.Println(s.collector.Logs[0].Contents[0].GetValue())
-	c.Assert(len(s.collector.Logs[0].Contents[0].GetValue()), check.Equals, 1+2+3)
-	fmt.Println("TestSplitedLine : ", s.collector.Logs)
-	nTimeout := processor.Process([]byte(splitedlog1), time.Minute)
+	value := s.collector.Logs[0].Contents[0].GetValue()
+	c.Assert(len(value), check.Equals, 1+2+3)
+	c.Assert(value, check.Equals, "122333")
+	c.Assert(helper.IsSafeString(str1, value), check.IsTrue)
+	c.Assert(helper.IsSafeString(str2, value), check.IsTrue)
+	c.Assert(helper.IsSafeString(str3, value), check.IsTrue)
+
+	nTimeout := processor.Process(splitedlog1Bytes, time.Minute)
 	c.Assert(nTimeout, check.Equals, len(splitedlog1))
 	c.Assert(len(s.collector.Logs), check.Equals, 2)
-	c.Assert(len(s.collector.Logs[1].Contents[0].GetValue()), check.Equals, 1)
+	value = s.collector.Logs[1].Contents[0].GetValue()
+	c.Assert(value, check.Equals, "1")
+	c.Assert(helper.IsSafeString(str1, value), check.IsTrue)
 }
 
 func (s *inputProcessorTestSuite) TestError(c *check.C) {
 	processor := NewDockerStdoutProcessor(nil, time.Duration(0), 0, 512*1024, true, true, &s.context, &s.collector, s.tag)
 
-	n := processor.Process([]byte(logErrorJSON), time.Duration(0))
+	bytes := []byte(logErrorJSON)
+	str := helper.ZeroCopyString(bytes)
+
+	n := processor.Process(bytes, time.Duration(0))
 	c.Assert(n, check.Equals, len(logErrorJSON))
 	c.Assert(s.collector.Logs[0].Contents[0].GetKey(), check.Equals, "content")
 	c.Assert(s.collector.Logs[0].Contents[1].GetKey(), check.Equals, "_time_")
 	c.Assert(s.collector.Logs[0].Contents[2].GetKey(), check.Equals, "_source_")
-	c.Assert(s.collector.Logs[0].Contents[0].GetValue(), check.Equals, logErrorJSON)
+	value := s.collector.Logs[0].Contents[0].GetValue()
+	c.Assert(value, check.Equals, logErrorJSON)
+	c.Assert(helper.IsSafeString(value, str), check.IsTrue)
 
-	n = processor.Process([]byte(logErrorJSON+"\n"), time.Duration(0))
+	bytes = []byte(logErrorJSON + "\n")
+	str = helper.ZeroCopyString(bytes)
+	n = processor.Process(bytes, time.Duration(0))
 	c.Assert(n, check.Equals, len(logErrorJSON)+1)
 	c.Assert(s.collector.Logs[1].Contents[0].GetKey(), check.Equals, "content")
 	c.Assert(s.collector.Logs[1].Contents[1].GetKey(), check.Equals, "_time_")
 	c.Assert(s.collector.Logs[1].Contents[2].GetKey(), check.Equals, "_source_")
-	c.Assert(s.collector.Logs[1].Contents[0].GetValue(), check.Equals, logErrorJSON)
+	value = s.collector.Logs[1].Contents[0].GetValue()
+	c.Assert(value, check.Equals, logErrorJSON)
+	c.Assert(helper.IsSafeString(value, str), check.IsTrue)
+
+	bytes = []byte(logErrorContainerd1)
+	str = helper.ZeroCopyString(bytes)
+	n = processor.Process(bytes, time.Duration(0))
+	c.Assert(n, check.Equals, len(logErrorContainerd1))
+	c.Assert(s.collector.Logs[2].Contents[0].GetKey(), check.Equals, "content")
+	c.Assert(s.collector.Logs[2].Contents[1].GetKey(), check.Equals, "_time_")
+	c.Assert(s.collector.Logs[2].Contents[2].GetKey(), check.Equals, "_source_")
+	value = s.collector.Logs[2].Contents[0].GetValue()
+	c.Assert(value, check.Equals, logErrorContainerd1)
+	c.Assert(helper.IsSafeString(value, str), check.IsTrue)
+
+	bytes = []byte(logErrorContainerd1 + "\n")
+	str = helper.ZeroCopyString(bytes)
+	n = processor.Process(bytes, time.Duration(0))
+	c.Assert(n, check.Equals, len(logErrorContainerd1)+1)
+	c.Assert(s.collector.Logs[3].Contents[0].GetKey(), check.Equals, "content")
+	c.Assert(s.collector.Logs[3].Contents[1].GetKey(), check.Equals, "_time_")
+	c.Assert(s.collector.Logs[3].Contents[2].GetKey(), check.Equals, "_source_")
+	value = s.collector.Logs[3].Contents[0].GetValue()
+	c.Assert(value, check.Equals, logErrorContainerd1)
+	c.Assert(helper.IsSafeString(value, str), check.IsTrue)
+
+	bytes = []byte(logErrorContainerd2 + "\n")
+	str = helper.ZeroCopyString(bytes)
+	n = processor.Process(bytes, time.Duration(0))
+	c.Assert(n, check.Equals, len(logErrorContainerd2)+1)
+	c.Assert(s.collector.Logs[4].Contents[0].GetKey(), check.Equals, "content")
+	c.Assert(s.collector.Logs[4].Contents[1].GetKey(), check.Equals, "_time_")
+	c.Assert(s.collector.Logs[4].Contents[2].GetKey(), check.Equals, "_source_")
+	value = s.collector.Logs[4].Contents[0].GetValue()
+	c.Assert(value, check.Equals, logErrorContainerd2)
+	c.Assert(helper.IsSafeString(value, str), check.IsTrue)
+
+	bytes = []byte(logErrorContainerd3 + "\n")
+	str = helper.ZeroCopyString(bytes)
+	n = processor.Process(bytes, time.Duration(0))
+	c.Assert(n, check.Equals, len(logErrorContainerd3)+1)
+	c.Assert(s.collector.Logs[5].Contents[0].GetKey(), check.Equals, "content")
+	c.Assert(s.collector.Logs[5].Contents[1].GetKey(), check.Equals, "_time_")
+	c.Assert(s.collector.Logs[5].Contents[2].GetKey(), check.Equals, "_source_")
+	value = s.collector.Logs[5].Contents[0].GetValue()
+	c.Assert(value, check.Equals, logErrorContainerd3)
+	c.Assert(helper.IsSafeString(value, str), check.IsTrue)
 }
 
 func (s *inputProcessorTestSuite) TestMultiLine(c *check.C) {
 	processor := NewDockerStdoutProcessor(regexp.MustCompile(`\d+-\d+-\d+.*`), time.Second, 12, 512*1024, true, true, &s.context, &s.collector, s.tag)
 
-	n := processor.Process([]byte(s.allLog), time.Second*time.Duration(0))
+	bytes := []byte(s.allLog)
+	str := helper.ZeroCopyString(bytes)
+	n := processor.Process(bytes, time.Second*time.Duration(0))
 	c.Assert(n, check.Equals, s.multiLineLen)
 
 	c.Assert(len(s.collector.Logs), check.Equals, 4)
-	c.Assert(s.collector.Logs[0].Contents[0].GetValue(), check.Equals, context1+"\n"+context2)
-	c.Assert(s.collector.Logs[1].Contents[0].GetValue(), check.Equals, context3)
-	c.Assert(s.collector.Logs[2].Contents[0].GetValue(), check.Equals, context4)
-	c.Assert(s.collector.Logs[3].Contents[0].GetValue(), check.Equals, context5)
+	value0 := s.collector.Logs[0].Contents[0].GetValue()
+	c.Assert(value0, check.Equals, context1+"\n"+context2)
+	c.Assert(helper.IsSafeString(value0, str), check.IsTrue)
+
+	value1 := s.collector.Logs[1].Contents[0].GetValue()
+	c.Assert(value1, check.Equals, context3)
+	c.Assert(helper.IsSafeString(value1, str), check.IsTrue)
+
+	value2 := s.collector.Logs[2].Contents[0].GetValue()
+	c.Assert(value2, check.Equals, context4)
+	c.Assert(helper.IsSafeString(value2, str), check.IsTrue)
+
+	value3 := s.collector.Logs[3].Contents[0].GetValue()
+	c.Assert(value3, check.Equals, context5)
+	c.Assert(helper.IsSafeString(value3, str), check.IsTrue)
 }
 
 func (s *inputProcessorTestSuite) TestMultiLineTimeout(c *check.C) {
 	processor := NewDockerStdoutProcessor(regexp.MustCompile(`\d+-\d+-\d+.*`), time.Second, 12, 512*1024, true, true, &s.context, &s.collector, s.tag)
 	lastLine := strings.LastIndexByte(s.allLog, '\n')
 	lastLine = strings.LastIndexByte(s.allLog[0:lastLine], '\n')
-	n := processor.Process([]byte(s.allLog[0:lastLine+1]), time.Second*time.Duration(0))
+	originBytes := []byte(s.allLog)
+	str := helper.ZeroCopyString(originBytes)
+	bytes := originBytes[0 : lastLine+1]
+
+	n := processor.Process(bytes, time.Second*time.Duration(0))
 	c.Assert(n, check.Equals, lastLine+1)
 	c.Assert(len(s.collector.Logs), check.Equals, 4)
-	c.Assert(s.collector.Logs[0].Contents[0].GetValue(), check.Equals, context1+"\n"+context2)
-	c.Assert(s.collector.Logs[1].Contents[0].GetValue(), check.Equals, context3)
-	c.Assert(s.collector.Logs[2].Contents[0].GetValue(), check.Equals, context4)
-	c.Assert(s.collector.Logs[3].Contents[0].GetValue(), check.Equals, context5)
-	n = processor.Process([]byte(s.allLog[lastLine+1:]), time.Second*time.Duration(0))
+
+	value0 := s.collector.Logs[0].Contents[0].GetValue()
+	c.Assert(value0, check.Equals, context1+"\n"+context2)
+	c.Assert(helper.IsSafeString(value0, str), check.IsTrue)
+
+	value1 := s.collector.Logs[1].Contents[0].GetValue()
+	c.Assert(value1, check.Equals, context3)
+	c.Assert(helper.IsSafeString(value1, str), check.IsTrue)
+
+	value2 := s.collector.Logs[2].Contents[0].GetValue()
+	c.Assert(value2, check.Equals, context4)
+	c.Assert(helper.IsSafeString(value2, str), check.IsTrue)
+
+	value3 := s.collector.Logs[3].Contents[0].GetValue()
+	c.Assert(value3, check.Equals, context5)
+	c.Assert(helper.IsSafeString(value3, str), check.IsTrue)
+
+	bytes = originBytes[lastLine+1:]
+	n = processor.Process(bytes, time.Second*time.Duration(0))
 	c.Assert(n, check.Equals, len(s.allLog)-lastLine-1)
 	c.Assert(len(s.collector.Logs), check.Equals, 4)
+
 	processor.Process(nil, time.Second*time.Duration(2))
 	c.Assert(len(s.collector.Logs), check.Equals, 5)
-	c.Assert(s.collector.Logs[4].Contents[0].GetValue(), check.Equals, context6)
+	value4 := s.collector.Logs[4].Contents[0].GetValue()
+	c.Assert(value4, check.Equals, context6)
+	c.Assert(helper.IsSafeString(value3, str), check.IsTrue)
 }
 
 func (s *inputProcessorTestSuite) TestMultiLineMaxLength(c *check.C) {
-	processor := NewDockerStdoutProcessor(regexp.MustCompile(`\d+-\d+-\d+.*`), time.Second, 12, 512*1024, true, true, &s.context, &s.collector, s.tag)
-
+	processor := NewDockerStdoutProcessor(regexp.MustCompile(`\d+-\d+-\d+.*`), time.Second, 12, 512*1056, true, true, &s.context, &s.collector, s.tag)
+	bytes := make([]byte, 3000)
+	str := helper.ZeroCopyString(bytes)
 	content := util.RandomString(1024)
 	for i := 0; i < 513; i++ {
 		realLine := fmt.Sprintf("2017-09-12T22:32:21.212861448Z stdout [%d] %s\n", i, content)
-		n := processor.Process([]byte(realLine), time.Second*time.Duration(0))
+		copy(bytes, realLine)
+		n := processor.Process(bytes[:len(realLine)], time.Second*time.Duration(0))
 		c.Assert(n, check.Equals, len(realLine))
+		if i != 512 {
+			c.Assert(len(s.collector.Logs), check.Equals, 0)
+			c.Assert(helper.IsSafeString(helper.ZeroCopyString(processor.lastLogs[i].Content), str), check.IsTrue)
+		}
 	}
 
 	c.Assert(len(s.collector.Logs), check.Equals, 1)
-	c.Assert(len(s.collector.Logs[0].Contents[0].GetValue()), check.Greater, 500*1000)
+	value := s.collector.Logs[0].Contents[0].GetValue()
+	c.Assert(len(value), check.Greater, 500*1000)
+	c.Assert(helper.IsSafeString(value, str), check.IsTrue)
 }
 
 func (s *inputProcessorTestSuite) TestMultiLineError(c *check.C) {
 	processor := NewDockerStdoutProcessor(regexp.MustCompile(`\d+-\d+-\d+.*`), time.Second, 12, 512*1024, true, true, &s.context, &s.collector, s.tag)
 
-	n := processor.Process([]byte(logErrorJSON), time.Second*time.Duration(0))
+	bytes := []byte(logErrorJSON)
+	str := helper.ZeroCopyString(bytes)
+	n := processor.Process(bytes, time.Second*time.Duration(0))
 	c.Assert(n, check.Equals, len(logErrorJSON))
 
 	c.Assert(len(s.collector.Logs), check.Equals, 1)
-	c.Assert(s.collector.Logs[0].Contents[0].GetValue(), check.Equals, logErrorJSON)
+	value := s.collector.Logs[0].Contents[0].GetValue()
+	c.Assert(value, check.Equals, logErrorJSON)
+	c.Assert(helper.IsSafeString(value, str), check.IsTrue)
 }
 
 func (s *inputProcessorTestSuite) TestBigLine(c *check.C) {
@@ -283,10 +417,9 @@ func (s *inputProcessorTestSuite) TestBigLine(c *check.C) {
 	processor := NewDockerStdoutProcessor(nil, time.Duration(0), 0, 512*1024, true, true, &s.context, &s.collector, s.tag)
 	for i := 0; i < 1000; i++ {
 		processor.Process(bigline, time.Second*time.Duration(0))
+		c.Assert(helper.IsSafeString(s.collector.Logs[i].Contents[0].GetValue(), helper.ZeroCopyString(bigline)), check.IsTrue)
 	}
-
 	c.Assert(len(s.collector.Logs), check.Equals, 1000)
-	fmt.Println(s.collector.Logs[0].Contents[0].GetValue())
 }
 
 func TestParseCRILine(t *testing.T) {
@@ -310,38 +443,82 @@ func TestParseCRILine(t *testing.T) {
 
 	// Single line.
 	{
+		bytes := make([]byte, 100)
+		str := helper.ZeroCopyString(bytes)
+
 		processor := NewDockerStdoutProcessor(nil, time.Duration(0),
 			0, 512*1024, true, true, &context, &collector, tags)
 
-		processor.Process([]byte("2021-07-13T16:32:21.212861448Z stdout F full line\n"), duration)
+		single := "2021-07-13T16:32:21.212861448Z stdout F full line\n"
+		copy(bytes, single)
+		processor.Process(bytes[:len(single)], duration)
 		require.Equal(t, len(collector.Logs), 1)
 		assertKeyValue(collector.Logs[0], "content", "full line")
+		assert.True(t, helper.IsSafeString(collector.Logs[0].Contents[0].GetValue(), str))
 
-		processor.Process([]byte("2021-07-13T16:32:21.212861448Z stdout P partial line:\n"), duration)
-		processor.Process([]byte("2021-07-13T16:32:21.212861448Z stdout F full line of partial line\n"), duration)
+		single = "2021-07-13T16:32:21.212861448Z stdout P partial line:\n"
+		copy(bytes, single)
+		processor.Process(bytes[:len(single)], duration)
+		assert.True(t, helper.IsSafeString(helper.ZeroCopyString(processor.lastLogs[0].Content), str))
+
+		single = "2021-07-13T16:32:21.212861448Z stdout F full line of partial line\n"
+		copy(bytes, single)
+		processor.Process(bytes[:len(single)], duration)
 		require.Equal(t, len(collector.Logs), 2)
 		assertKeyValue(collector.Logs[1], "content", "partial line:full line of partial line")
+		assert.True(t, helper.IsSafeString(collector.Logs[1].Contents[0].GetValue(), str))
 	}
 
 	// Multiple lines.
 	collector.Logs = nil
 	{
+		bytes := make([]byte, 100)
+		str := helper.ZeroCopyString(bytes)
+
 		timeoutDuration := time.Duration(1) * time.Second
 		processor := NewDockerStdoutProcessor(
 			regexp.MustCompile(`\d+-\d+-\d+.*`), timeoutDuration,
 			10, 512*1024, true, true, &context, &collector, tags)
 
-		processor.Process([]byte("2021-07-13T16:32:21.212861448Z stdout F 2021-07-13 full line line 1\n"), duration)
-		processor.Process([]byte("2021-07-13T16:32:21.212861448Z stdout F   full line line end\n"), duration)
-		require.Equal(t, len(collector.Logs), 0)
+		single := "2021-07-13T16:32:21.212861448Z stdout F 2021-07-13 full line line 1\n"
+		copy(bytes, single)
 
-		processor.Process([]byte("2021-07-13T16:32:21.212861448Z stdout P 2021-07-13 partial line line 1 partial\n"), duration)
+		processor.Process(bytes[:len(single)], duration)
+		assert.True(t, helper.IsSafeString(helper.ZeroCopyString(processor.lastLogs[0].Content), str))
+
+		single = "2021-07-13T16:32:21.212861448Z stdout F   full line line end\n"
+		copy(bytes, single)
+		processor.Process(bytes[:len(single)], duration)
+		require.Equal(t, len(collector.Logs), 0)
+		assert.True(t, helper.IsSafeString(helper.ZeroCopyString(processor.lastLogs[1].Content), str))
+
+		single = "2021-07-13T16:32:21.212861448Z stdout P 2021-07-13 partial line line 1 partial\n"
+		copy(bytes, single)
+		processor.Process(bytes[:len(single)], duration)
 		require.Equal(t, len(collector.Logs), 1)
 		assertKeyValue(collector.Logs[0], "content", "2021-07-13 full line line 1\n  full line line end")
+		assert.True(t, helper.IsSafeString(collector.Logs[0].Contents[0].GetValue(), str))
+		assert.Equal(t, len(processor.lastLogs), 1)
+		assert.True(t, helper.IsSafeString(helper.ZeroCopyString(processor.lastLogs[0].Content), str))
 
-		processor.Process([]byte("2021-07-13T16:32:21.212861448Z stdout F   partial line line 1 full\n"), duration)
-		processor.Process([]byte("2021-07-13T16:32:21.212861448Z stdout P   partial line line 2 partial\n"), duration)
-		processor.Process([]byte("2021-07-13T16:32:21.212861448Z stdout F   partial line line 2 full\n"), duration)
+		single = "2021-07-13T16:32:21.212861448Z stdout F   partial line line 1 full\n"
+		copy(bytes, single)
+		processor.Process(bytes[:len(single)], duration)
+		assert.Equal(t, len(processor.lastLogs), 2)
+		assert.True(t, helper.IsSafeString(helper.ZeroCopyString(processor.lastLogs[1].Content), str))
+
+		single = "2021-07-13T16:32:21.212861448Z stdout P   partial line line 2 partial\n"
+		copy(bytes, single)
+		processor.Process(bytes[:len(single)], duration)
+		assert.Equal(t, len(processor.lastLogs), 3)
+		assert.True(t, helper.IsSafeString(helper.ZeroCopyString(processor.lastLogs[2].Content), str))
+
+		single = "2021-07-13T16:32:21.212861448Z stdout F   partial line line 2 full\n"
+		copy(bytes, single)
+		processor.Process(bytes[:len(single)], duration)
+		assert.Equal(t, len(processor.lastLogs), 4)
+		assert.True(t, helper.IsSafeString(helper.ZeroCopyString(processor.lastLogs[3].Content), str))
+
 		require.Equal(t, len(collector.Logs), 1)
 
 		processor.Process([]byte(""), timeoutDuration+time.Microsecond)
@@ -349,5 +526,96 @@ func TestParseCRILine(t *testing.T) {
 		assertKeyValue(collector.Logs[1], "content",
 			"2021-07-13 partial line line 1 partial  partial line line 1 full\n"+
 				"  partial line line 2 partial  partial line line 2 full")
+		assert.True(t, helper.IsSafeString(collector.Logs[1].Contents[0].GetValue(), str))
 	}
+}
+
+func TestSingleLineChangeBlock(t *testing.T) {
+	var context helper.LocalContext
+	var collector helper.LocalCollector
+	tags := map[string]string{
+		"container":    "test",
+		"container_id": "id",
+	}
+	processor := NewDockerStdoutProcessor(nil, time.Duration(0),
+		0, 512*1024, true, true, &context, &collector, tags)
+
+	// valid single line when change block
+	{
+		str := "{\"log\":\"1:M 09 Nov 13:27:36.276 # User requested shutdown...\\n\",\"stream\":\"stdout\", \"time\":\"2018-05-16T06:28:41.2195434Z\"}\n"
+		dockerSingleLine := helper.ZeroCopySlice(str)
+		assert.Equal(t, processor.Process(dockerSingleLine, time.Hour), len(dockerSingleLine))
+		value := collector.Logs[0].Contents[0].GetValue()
+		assert.Equal(t, value, "1:M 09 Nov 13:27:36.276 # User requested shutdown...")
+		assert.True(t, helper.IsSafeString(str, value))
+
+		assert.Equal(t, collector.Logs[0].Contents[0].GetValue(), "1:M 09 Nov 13:27:36.276 # User requested shutdown...")
+
+		str = "2021-07-13T16:32:21.212861448Z stdout F full line line end\n"
+		contianerdSingleLine := helper.ZeroCopySlice(str)
+		assert.Equal(t, processor.Process(contianerdSingleLine, time.Hour), len(contianerdSingleLine))
+		value = collector.Logs[1].Contents[0].GetValue()
+		assert.Equal(t, value, "full line line end")
+		assert.True(t, helper.IsSafeString(str, value))
+		collector.Logs = collector.Logs[:0]
+	}
+
+	// valid split line when change block
+	{
+		part1 := []byte("2021-07-13T16:32:21.212861448Z stdout P partial line:\n")
+		str := helper.ZeroCopyString(part1)
+
+		assert.Equal(t, processor.Process(part1, 0), len(part1))
+		assert.Equal(t, len(collector.Logs), 0)
+		part1[38] = part1[38] - 10
+		assert.Equal(t, processor.Process(part1, 0), len(part1))
+		assert.Equal(t, len(collector.Logs), 1)
+		value := collector.Logs[0].Contents[0].GetValue()
+		assert.Equal(t, value, "partial line:partial line:")
+		assert.True(t, helper.IsSafeString(str, value))
+		collector.Logs = collector.Logs[:0]
+	}
+
+	// valid containerd multi line when change block
+	{
+		processor := NewDockerStdoutProcessor(regexp.MustCompile(`^\d+-\d+-\d+.*`), time.Duration(0),
+			10, 512*1024, true, true, &context, &collector, tags)
+		part1 := []byte("2017-09-12T22:32:21.212861448Z stderr 2017-09-12 22:32:21.212 [INFO][88] exception 1: \n")
+		str := helper.ZeroCopyString(part1)
+
+		assert.Equal(t, processor.Process(part1, 0), len(part1))
+		assert.Equal(t, len(collector.Logs), 0)
+		part1[38] += 40
+		assert.Equal(t, processor.Process(part1, 0), len(part1))
+		assert.Equal(t, len(collector.Logs), 0)
+		part1[38] -= 40
+		assert.Equal(t, processor.Process(part1, 0), len(part1))
+		assert.Equal(t, len(collector.Logs), 1)
+		value := collector.Logs[0].Contents[0].GetValue()
+		assert.Equal(t, value, "2017-09-12 22:32:21.212 [INFO][88] exception 1: \nZ017-09-12 22:32:21.212 [INFO][88] exception 1: ")
+		assert.True(t, helper.IsSafeString(str, value))
+		collector.Logs = collector.Logs[:0]
+	}
+
+	// valid docker multi line when change block
+	{
+		processor := NewDockerStdoutProcessor(regexp.MustCompile(`^\d+-\d+-\d+.*`), time.Duration(0),
+			10, 512*1024, true, true, &context, &collector, tags)
+		part1 := []byte("{\"log\":\"2017-09-12 22:32:21.212 13:27:36.276 # User requested shutdown...\\n\",\"stream\":\"stdout\", \"time\":\"2018-05-16T06:28:41.2195434Z\"}\n")
+		str := helper.ZeroCopyString(part1)
+
+		assert.Equal(t, processor.Process(part1, 0), len(part1))
+		assert.Equal(t, len(collector.Logs), 0)
+		part1[8] += 40
+		assert.Equal(t, processor.Process(part1, 0), len(part1))
+		assert.Equal(t, len(collector.Logs), 0)
+		part1[8] -= 40
+		assert.Equal(t, processor.Process(part1, 0), len(part1))
+		assert.Equal(t, len(collector.Logs), 1)
+		value := collector.Logs[0].Contents[0].GetValue()
+		assert.Equal(t, value, "2017-09-12 22:32:21.212 13:27:36.276 # User requested shutdown...\nZ017-09-12 22:32:21.212 13:27:36.276 # User requested shutdown...")
+		assert.True(t, helper.IsSafeString(str, value))
+		collector.Logs = collector.Logs[:0]
+	}
+
 }